control-service: reduce the k8s job deadline to 12 hours
Currently, Kubernetes jobs used for data job executions are configred by default
to run for up to 36 hours before being terminated.
This is a long time for a single job execution. After examining the potential
impact of reducing this value on existing adoptions of the Versatile Data Kit,
I have decided to modify the deadline to use a more meaningful default of 12 hours.

Testing done: deployed a new data job and verified that its activeDeadlineSeconds is
now 12 hours.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>