control-service: Turn-off debug mode for builder script
Builder job pods run a shell script, which builds and uploads the
data job image. Currently, it runs in debug mode which is not needed
in production environments.

This change turns off debug mode, and logs only required and non-sensitive
steps.

Testing Done: Uploaded new builder job image to docker hub, and ran a local
instance of the Control Service to verify that only necessary data is logged.

Signed-off-by: Andon Andonov <andonova@vmware.com>