control-service: add last execution info to the data_job table
As part of preparation for the implementation of the ability to filter
and sort data jobs by their last execution status, time, and duration,
add the respective columns to the data_job table.

Testing done: Start the service and verify that the columns are created
in the data_job table.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>