control-service: add last execution info to the data_job table (#480)
As part of the preparation for the implementation of the ability to filter
and sort data jobs by their last execution status, time, and duration,
add the respective columns to the data_job table.

The H2 database used when running tests fails if the three new
columns are added all at once. That's why the script contains three
queries to insert the columns one by one.

Testing done: Start the service and verify that the columns are created
in the data_job table. Also, run the service locally against an in-memory
H2 database and verified that the script is run successfully.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>