control-service: Classify OOM as User Errors (#479)
Currently, if a data job execution exceeds the allowed memory
quota, the pod of the execution is immediately killed by
Kubernetes and restarted. Which in turn usually results in a
similar failure. This causes a Platform Error to be raised, whereas
it should be classified as a User Error.

This change re-classifies such errors as User Errors.

Testing Done: Added unit test.

Signed-off-by: Andon Andonov <andonova@vmware.com>