control-service: Update the GraphQL model with the last execution
Extended the GraphQL and service models to reflect the properties that were added
as part of the implementation of the ability to sort and filter data jobs by their
last execution status, time, and duration.

Wire the two models so that the newly added properties can be queried via GraphQL.

Testing done: Run the service locally and verified that the new properties are
returned when requested.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>