[vdk-core]: Post-ingestion specification
With the increase in the number of use-cases for ingestion,
there arises the need for some users to do post-processing
of ingestion data (e.g, to collect ingestion telemetry, etc.).

This change aims at introducing a new interface for plugin developers
to use when they want to do further processing of payloads after the
ingestion operation is complete, or to send telemetry data based on the
ingestion operation.

Testing Done: Unit test

Signed-off-by: Andon Andonov <andonova@vmware.com>