vdk-core: Fix showing None default values
When checking whether a default value should be shown in config-help,
vdk-core would set the default value to None in the case where it
shouldn't be shown. However, this has the unintended consequence of
not showing default values which are explicitly set to None. This
change fixes that by creating a new constant string and checking for
its value instead of None.

Testing done: ran config-help and observed the shown None default value,
tested for both values of show_default_value parameter when adding
config variables

Signed-off-by: gageorgiev <gageorgiev@vmware.com>