control-service: Add logging for keytab secret creation
why: A user identified a problem with a data job deployment which
was related to keytab secret creation. The data job was not executing
due to missing secret keytab. Logs weren't enough to debug the
issue, so we need to add additional logs.

what: Added logging around the secret creation logic.

testing: trivial.

Signed-off-by: mrMoZ1 mzhivkov@vmware.com