control-service: fix OOM classification as User error
Recently we implemented a fix to classify job executions that run OOM
as User errors. For this purpose, we observe the status of the
executing pod. When the K8s job first terminates, its pod is inspected
and the termination status is correctly set to User error. However,
during subsequent updates, the pod may disappear and this is
interpreted as Platform error.

With this commit, we prevent changes to the termination status of an
execution if the status is final (i.e. not None).

Testing done: new unit tests

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>