control-service: address review comments
In addition to addressing review comments, this commit also changes
how the last execution status is populated. Previously, the status
was updated from the most recently updated execution of a data job.
However, there is a possibility that the most recently updated execution
is no the last one that completed. With this commit the last execution
status is updated from the last completed execution.

Testing done: integration tests

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>