control-service: graphQL api - add job success/fail rate
why: Users requested the ability to produce reports
based on data job execution failure/success numbers.
This is a part of the graphQL epic. Closes #424

what: Integrated already present functionality so that
fail and success counts are included in the graphql
response, if the fields are requested in the call. Field
query and population happens during the post pagination
phase.

testing: Extended unit test to cover new functionality.

Signed-off-by: Momchil Zhivkov mzhivkov@vmware.com