control-service: set build job impagePullPolicity to default to IfNotPresent
We rarely change the version (tag) of the builder job image. There's not
point of trying to pull it every time , so this is setting it to
IfNotPresent. This way we can limit the chance of hitting docker limits.

It's made configurable using
`datajobs.deployment.builder.imagePullPolicy` system property in case in
some environments it is necessary.

Testing Done: integration test passed

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>