vdk-greenplum: Fix ingestion issue
The ingestion query execution was not being committed, which meant
that querying for the ingested data returned a newline character.
This issue might be related to the newest psycopg2 release, which
was made 11 days ago, although that's highly unlikely as these API's
tend to be fairly ossified and searching for newly reported issues
gave me nothing.

Testing done: tested against local Greenplum instance, unit tests

Signed-off-by: gageorgiev <gageorgiev@vmware.com>