control-service: change how termination status is updated
Up until now the termination status of a data job is calculated
independently from how executions are persisted, which sometimes
could cause problems. For example, because of this, when we
receive an event for an older execution it is not possible
to recognize this and the termination status is updated with it
regardless, causing it to become incorrect.

This commit changes how termination status is updated. Now the
event is processed first and the execution table is updated with it.
Then the termination status is updated with the last execution of
the data job, regardless if the event was related to it or not.

Testing done: unit and integration tests pass

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>