vdk-heartbeat: make sure op_id is passed everywhere
OpId is used to track the same operation across multiple services, CLIs,
data (ala open tracing style). It helps troubleshooting efficiently -
just search for the op id of the failed process we are troubleshooitng
and you have end to end data flow picture.

Testing Done: vdk-heartbeat -f local.ini  and verified op id is printed
in the local logs and in the logs of the Control service (and passed to
the data job execution)

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>