vdk-core: Move new version printout to the end of the log
If VDK or some plugin has a newer version, VDK automatically
prints out a message stating this and prompting the user to
update. However, this is printed out at the start of the log,
and can be easily missed.
This change moves the printout to the end of the log.

Testing done: verified printout appears at the end

Signed-off-by: Gabriel Georgiev <gageorgiev@vmware.com>