vdk-heartbeat: Make vdk version configurable (#576)
We want to be able to run vdk-heartbeat as a pre-release
step in out CI/CD pipelines. This requires running it with a
specific vdk version.

Add VDK_HEARTBEAT_DEPLOY_JOB_VDK_VERSION configuration variable
in vdk-heartbeat. If set, it updates the job deployment to use
the specified vdk version.

Testing done: run vdk-heartbeat with a specified version against
a local control service and observe that this version is used
in the execution k8s pod.

Signed-off-by: Yana Zhivkova yzhivkova@vmware.com