vdk-impala: Error handling and recovery support
Database connection plugins should be able to handle and recover
from errors in the best way possible.
This change introduces error handling and error recovery to the
vdk-impala plugin. After this change, in the case of an error,
vdk-impala will analyze the error and either retry the query if
possible, execute any necessary additional operation, or print
a descriptive message.

Signed-off-by: Gabriel Georgiev <gageorgiev@vmware.com>