vdk-core: provide needed error details for vdk run
There was some corner cases that error message was missing - if job
failed with exception for example during initialize or finalize phases.

Also made sure job finishes with summary of job execution so it's easier
for users to see what happened.
Also if job step failed - it was printing the excepiton twice (at step
level and job level) which is just cluttering the logs

Tesitng Done: unit tests.

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>