vdk-test-utils: Move connection test util func to vdk-test-utils
The testing of the connection builtin plugin requires a util func
which was currently stored inside a single file in the testing dir.
This is suboptimal, as testing utility functions should be included
in the vdk-test-utils plugin.
Additionally, the aforementioned function was amended to allow the
passing of a decoration_operation_callback as an optional parameter.

Signed-off-by: Gabriel Georgiev <gageorgiev@vmware.com>