vdk-core: Fix inclusion of exception message in log call (#601)
Certain log calls would include the exception without specifying a
substitution string inside in the logged message for it. This lead
to an error when that log call is executed. This change fixes that
by replacing the log message with an f-string and includes the
exception in it.
Additionally, it replaces certain log calls in the code base with
f-strings as they are the recommended way of formatting strings.

Signed-off-by: Gabriel Georgiev <gageorgiev@vmware.com>