control-service: graphQl add sorting by notifications
why: Users requested the ability to sort DataJobs
by notifications. This is a part of the graphql epic.

what: Added the ability to sort by notifications enabled
or not. We check if one of the contacts is populated and
if it is, we threat it as a notification enabled. The DataJobs
will then get sorted by notification enabled status

testing: Added unit tests.

Signed-off-by: Momchil Zhivkov mzhivkov@vmware.com