vdk-impala: Move time.sleep fix to conftest
The vdk-impala functional tests would require a time.sleep statement
at their start due to an issue where the Impala would return a response
with exit code 0, but the service would not yet be available, which
would cause tests to fail.
This change moves that time.sleep call to the pytest fixture function
for two reasons: it logically makes more sense for it to be there, and
this way it will only ever be invoked once, instead of at the start of
every single functional test.

Signed-off-by: Gabriel Georgiev <gageorgiev@vmware.com>