vdk-core: Address Feedback for Introduce Ingestion Payload processing sequence interfaces
As part of the improvements to the Ingestion API, it is necessary to make it
possible for multiple ingestion plugins to be chained and provide capability for
complex payload processing.

This change introduces new interface for the IngesterPlugin to facilitate payload
processing without ingesting, and a new environment variable in the ingestion configuration
to allow clients to specify the sequence in which ingestion plugins get executed.

Testing Done: Added tests that would be used when the implementation of the interfaces is complete.

Signed-off-by: Andon Andonov <andonova@vmware.com>