vdk-core: Publish release candidate distribution and image
We want to avoid releasing a broken vdk-core distribution as
much as possible. This is why we are working on adding a
pre-release step in vdk-core's CI/CD which would run vdk-heartbeat
with the new version (release candidate) before actually releasing
it.

Create release candidate vdk-core distribution and base image.
Patch version of the release-candidate distribution:
${CI_PIPELINE_ID}dev.${CI_PIPELINE_IID} - corresponding to
https://packaging.python.org/en/latest/guides/distributing-packages-using-setuptools/#pre-release-versioning

Testing Done:
- released a test distribution to
https://test.pypi.org/legacy/ by manually executing the steps
from vdk-core-release-candidate CI/CD job. Then installed it and ran
"vdk version" to check that it is as expected.
- CI/CD pipelines

Signed-off-by: Yana Zhivkova <yzhivkova@vmware.com>