control-service: improve logging during job monitoring
Currently, when an exception happens during k8s job watching,
we swallow the exception and log only the exception message.
This prevents us from pinpointing the exact location of the
exception and makes troubleshooting it harder.

This commit also logs the exceptions' stack trace.

Testing done: verified that the stack trace is logged on exception.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>