control-service: add logging when updating last execution
We got reports that the last execution status of a data job
can get out-of-sync with the status persisted in the data_job_execution
table. This commit adds logging to the method that updates the last
execution information of a data job so that it is easier to troubleshoot.

Testing done: unit tests

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>