quickstart-vdk: Run vdk-heartbeat before release
Currently, plugin versions are not fixed in quickstart-vdk
and when we upgrade, we can accidentally break it. That is why it is
necessary to run some verification before we release it.

Run vdk-heartbeat before releasing quickstart-vdk (during pre-release stage).
Create release-candidate package and image of quickstart-vdk and
run vdk-heartbeat with them during pre-release stage of
each quickstart-vdk dependency CI/CD.

Testing Done: test CI/CD pipeline. Also, check that the release-candidate
image of quickstart-vdk is run on the vdk-heartbeat job k8s pod.

Signed-off-by: Yana Zhivkova <yzhivkova@vmware.com>