vdk-kerberos-auth: introduce a new plugin (#629)
* vdk-kerberos-auth: add initial implementation

This commit introduces a new plugin - vdk-kerberos-auth,
which serves to provide Kerberos authentication.

The implementation is not yet complete - there will be
subsequent commits that will build upon this one. For
example HTTP requests are not yet authenticated.

Testing done: manually started the vdk with the new plugin
installed and verified that authorization is performed based
on the configuration (e.g. KRB_AUTH).

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>