vdk-kerberos-auth: provide variable to specify krb5.conf configuration
At this point the Kerberos authentication is always using a generic
krb5.conf file built into the plugin. This commits enables specifying
the path to the krb5.conf file via configuration (`KRB5_CONF_FILENAME`).
If the configuration is not set, it defaults to the built in file.

Testing done: manually runing the `run` command with and without the
new configuration and observing that the appropriate krb5.conf file is
being used.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>