versatile-data-kit: Establish release process
This change establishes a release process for the entire VDK repo.
The repo now has a global version.txt file, which, when changed,
caused CICD release steps to trigger for all components, with
versions set to the version in the global file. The automatic
releases made are now set to release as dev builds, and will not
be installed by default unless `pip install` is ran with the
`--pre` flag enabled.
The exact process is described in the CONTRIBUTING.md file.

Signed-off-by: Gabriel Georgiev <gageorgiev@vmware.com>