vdk-trino: trino ingest to handle type casting and missing values
Fix trino ingestion to handle different types of more gracefully.
It will make sure the the data type (passed in python) is correctly set
based on the data type in the destination table.

Handle missing values (e.g if a value for column X is missing , it will
set it to None / NULL insted of failing

Testing Done: extended functional tests to cover the new scenarios

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>