[vdk-plugins] Update ingestion interfaces used in plugins
As part of https://github.com/vmware/versatile-data-kit/pull/682, the
signature of the ingest_payload() method was updated to accept a metadata
object as part of the ingestion flow.

Once the interface change is adopted in vdk-core, this would cause data jobs to
fail, due to missing keyword argument "metadata".

This change updates the signatures of the ingest_payload() methods, implemented
in the vdk-plugins that provide ingestion capabilities. This is done in order to
avoid the abovementioned error.

Testing Done: Not required, as no plugins currently process the metadata.

Signed-off-by: Andon Andonov <andonova@vmware.com>