control-service: add timeouts to shedlock's database operations
We have experienced cases when operations against the database appear to hang
indefinitely (usually after a connectivity issue with the database). As a result
tasks that attempt to obtain a distributed lock, fail to start. This is
particularly problematic for important tasks, such as the data job watching.

This commit attempts to resolve this by specifying a query timeout to the
Jdbc template used by shedlock to operate with the database.

Link: https://dzone.com/articles/threads-stuck-in-javanetsocketinputstreamsocketrea

Testing done: started the service locally and veried that the locks are
operating normally with thevtimeout.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>