vdk examples: life-expectancy minor fixes to work end-to-end
The requirements.txt contained modules which are part of standard
library so it was failing to install them.

Set flag ingester_wait_to_finish_after_every_send since ingester is
asynchrounous and does not block after each method. Which causes
processing step to be executed before data is ingested

Testing Done: ran the job with the flag set and it was working correctly
(all data is processed)

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>