vdk-core: add flag to enable synchronous/blockng ingestion (#698)
We need to enable data jobs which want to combine ingestion and
processing steps in a single job. Currently because ingestion is
synchrounous if we add ingestion step and processing step afterwards ,
the processing step would not use the latest data.

The life-expectancy job
https://github.com/vmware/versatile-data-kit/tree/main/examples/life-expectancy
- is an example job.
It first ingests some data into Trino and then cleans in subsequent
steps. What happen oftne is that the processing steps work with empty
tables because the data has not been inserted/ingested yet.

We are adding new flag: INGESTER_WAIT_TO_FINISH_AFTER_EVERY_SEND which
when send will cause all calls to send_tabular_data_for_ingestion and
send_object_for_ingestion to block until the payloads is actually
processsed and send (to the plugin)

Testing Done: added unit tests which cover those scenarios.

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>