vdk-control-cli: Expand missing resource error message
In the case when a user creates a job in the control service with
a given team name, then changes the team name in the job's config
file, the message printed is not clear enough.
The message was expanded to inform the user of the aforementioned
possible issue. This approach was chosen over checking for this
issue specifically, as that approach would have required changes
to both vdk-control-cli and the control-service and would have
required much longer development time.

Closes #605

Testing done: following reproduction steps found in #605 and
verified message is printed

Signed-off-by: Gabriel Georgiev <gageorgiev@vmware.com>