control-service: add job-base-image folder and CI step/job
Currently the building of the data job base image was not done by our
CICD and was left in the older builder folder (vdk_job_builder). This
means that if there were changes to teh base image we'd need to rebuild
it manually (and not part of CI) which is not good.

After this change I'd completely delete vdk_job_builder which no longer
needed for anything.

Testing Done: this PR (the commit was edited to allow to run on PR , at
the end the CI job will run only in main)

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>