Online Exhibition Scenario
The Online Exhibition scenario implements a VDK use case using the REST API. It connects to the Europeana REST API and populates a database with the output of the query. The query asks for all the Van Gogh's artworks.
The scenario also implements an app in streamlit that shows a preview of all the downloaded artworks.

Tests have been done manually.

This pull request adds a directory named online-exhibition under the examples directory.

Signed-off-by: Angelica Lo Duca angelica [dot] loduca [at] gmail [dot] com