vdk-control-cli: Parse contacts with both comma "," as delimiter as well (#719)
Until now we required contacts be separated by ";" semi-colon. This was
error prone as a lot of users are using "," comma and we are not really
preventing them so mails were written as "a@abv.bg,b@dir.bg" as if it is
a single email. T This also meant that people were not getting
notifications.

The fix is to add "," comma as allowed delimiter as a first step.
In a subsequent change we can add a email validator (like
https://pypi.org/project/email-validator)

Testing Done: unit tests including new ones

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>