job-builder: add amazon-ecr-credential-helper to the job builder (#723)
* job-builder: add amazon-ecr-credential-helper to the job builder

The rootless job builder is based on an image that does not contain
the amazon-ecr-credential-helper utility. This utility is necessary
to authenticate against ECR when pushing data job images.

This commit adds the amazon-ecr-credential-helper utility to the job-builder
image by pulling it and building it locally from source. Maybe there is
a better way of provisioning it but I could not find it out.

Testing done: manually build and push a job-builder image with the change,
then create a k8s job with this image and run it to verify that it manages
to successfully push data job images to ECR.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>

* increment the builder version

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>

Co-authored-by: Miroslav Ivanov <40535952+mivanov1988@users.noreply.github.com>