control-service: fix log link to set endTime always
In case the job is not finished, there's no end time (logically) but
VMware Log Insight links do not work if no endTime is set. That might be
problem for other engines.

It's better to set it to current time instead of leaving empty.

I was wondering if I should set it to 1 hour in the future. But uses can
use endOffset option to change this. And there's chance that some Log
services may not work well with "future" times.

Testing Done: unit tests

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>