vdk-heartbeat: Handle execution end_time not string
Currently, if for some reason the `end_time` field of the execution object
is not a string (for example, the value is `None`), the `datetime.fromisoformat()`
method throws an exception and fails the heartbeat test.

This change introduces an explicit conversion of the value to string and adds
logging for what the error is along with indication that the heartbeat will fail.

Testing Done: Unit tests added.

Signed-off-by: Andon Andonov <andonova@vmware.com>