vdk-trino: fix ingesting value with bool type failing
vdk-trino ingestion was failing if `send_object_for_ingestion` is passed
value with type bool (as it was trying to cast it as if it's stirng.
Error was

"We could not convert the value to that type. Error is bool cast accept
only True/true/False/false values."

Testing Done: unit tests covered the new case

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>