vdk-kerberos-auth: add KerberosClient for authenticating API calls
We are adding support for Kerberos authentication in the Control REST
API . This is extending the support for the client to be able to make
authentication requests with kerberos.

This is part 1 adding KerberosClient "API" for the purpose of clients
using it to get Kerberos authentication header.

In a follow-up MR I will adopt it in vdk-control-cli

Testing Done: automated tests

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>