vdk-control-cli: add support for kerberos authentication
We are adding support for Kerberos authentication in the Control REST
API . This is extending the support for the client to be able to make
authentication requests with kerberos.

The implementation builds on previous by trying to stay smart.

1. If valid Oauth2 token is detected and Oauth2 auth is not disabled
(e.g by vdk logout) it will be used
2. If No valid Oauth2 token is detected and kerberos ticket is detected
(this happen if KeberosTicket can generate auth_header) and kerberos is
not disabled (by flag VDK_CONTROL_KERBEROS_AUTHENTICATION_ENABLED)
3. If none is detected - then no Authenticaiton is used.

Testing Done: automated tests

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>