[helm-chart]: add kerberos auth properties to helm chart (#764)
* [helm-chart]: add kerberos auth properties to helm chart

what: Added new properties to the helm chart as requested here:
#755 (comment)
The new properties expose kerberos authentication configuration properties
in the helm chart which are propagated to control-service.

why: We are introducing a new KerberosAuthProvider in control-service
and we need to expose configuration properties through the helm chart.

testing: Deployed on local minikube cluster. Made sure secret is created
and populated with correct files. Inspected control-service pods, made sure
the new env variables are present and that secrets exist.

Signed-off-by: Momchil Zhivkov mzhivkov@vmware.com

Co-authored-by: ivakoleva <ikoleva@vmware.com>