VEP-554: Extend Security and Permissions section
As part of the development of the Versatile Data Kit Airflow Provider, we need to
ensure that all connections to the Control Service are properly authenticated. For
this to happen, we need to define the authentication mechanisms and flows that would
be used.

This change extends the `Security and Permissions` section of the enhancement proposal
to include more details on how authentication would be facilitated in the Airflow Provider.

Testing Done: Documentation change, no testing needed.

Signed-off-by: Andon Andonov <andonova@vmware.com>