vdk-impala: unify names of templates betwen trino and impala
The "load/dimension/scd2" is simple implementation of SCD2 which is less
optimial and flexible than "load/versioned" which is the recommendd one
to use.

The "versioned" template correspond to scd2 in vdk-trino plugin for
example (vdk-trino/src/vdk/plugin/trino/templates/load/dimension/scd2) .
And there there's no scd2_simple implementation at all.

So this PR is changing the name of "load/dimension/scd2" to
"scd2_simple", It's keeping an alias (alternative) name of
"load/dimension/scd2" for backwards compatible reasons.

It's also setting as main name of load/versioned to be "scd2" (which is
the one used in vdk-trino)

Also in vdk-trino Periodic snapshot template is called
"periodic_snapshot" so renaming it to match.

Testing Done: automated tests

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>