airflow-provider-vdk: VDKSensor initial structure
This change lays out the initial structure for the VDK asynchronous
operator, a.k.a. the VDKSensor. The purpose of this sensor is to
routinely poke the VDK control-service and check the status of a
particular job, so that its completion/failure can trigger some
activity in an Airflow DAG, i.e. trigger another job.

Signed-off-by: Gabriel Georgiev <gageorgiev@vmware.com>