vdk-control-api-auth: Add core auth logic
As part of the work on extracting the authentication logic of vdk-control-cli
and moving it into a stand-alone library, we need to have a generalized
approach to authentication which is compliant with RFC-6749 while being able
to be integrated with the existing components of Versatile Data Kit, which
require authentication.

This change introduces the core authentication logic, which will be used and
extended to facilitate the different authentication workflows (api/refresh token
flow, pkce flow, etc.).

Testing Done: Unit tests added.

Signed-off-by: Andon Andonov <andonova@vmware.com>